Move error handler to api file
